Abstract version_base out of version
