+++ src/ctl/plugins/version.py
@@ -5,33 +5,21 @@ Plugin that allows you to handle repository versioning
 import argparse
 import os
 
-import toml
-
 import confu.schema
 
 import ctl
 from ctl.auth import expose
 from ctl.docs import pymdgen_confu_types
-from ctl.exceptions import OperationNotExposed, PluginOperationStopped, UsageError
-from ctl.plugins import ExecutablePlugin
-from ctl.plugins.changelog import ChangelogVersionMissing
-from ctl.plugins.changelog import temporary_plugin as temporary_changelog_plugin
-from ctl.plugins.git import temporary_plugin as temporary_git_plugin
-from ctl.plugins.repository import RepositoryPlugin
+from ctl.exceptions import OperationNotExposed, UsageError
+from ctl.plugins.version_base import VersionBasePlugin, VersionBasePluginConfig
 from ctl.util.versioning import bump_semantic, version_string
 
 
 @pymdgen_confu_types()
-class VersionPluginConfig(confu.schema.Schema):
+class VersionPluginConfig(VersionBasePluginConfig):
     """
     Configuration schema for `VersionPlugin`
     """
-
-    repository = confu.schema.Str(
-        help="name of repository type plugin or path " "to a repository checkout",
-        default=None,
-        cli=False,
-    )
 
     branch_dev = confu.schema.Str(
         default="master",
@@ -43,45 +31,15 @@ class VersionPluginConfig(confu.schema.Schema):
         help="the breanch to merge to when " "the --merge-release flag is present",
     )
 
-    changelog_validate = confu.schema.Bool(
-        default=True,
-        help="If a changelog data file (CHANGELOG.yaml) exists, validate before tagging",
-    )
-
 
 @ctl.plugin.register("version")
-class VersionPlugin(ExecutablePlugin):
+class VersionPlugin(VersionBasePlugin):
     """
     manage repository versioning
     """
 
-    class ConfigSchema(ExecutablePlugin.ConfigSchema):
+    class ConfigSchema(VersionBasePlugin.ConfigSchema):
         config = VersionPluginConfig()
-
-    @classmethod
-    def add_repo_argument(cls, parser, plugin_config):
-        """
-        The `repository` cli parameter needs to be available
-        on all operations. However since it is an optional
-        positional parameter that cames at the end using shared
-        parsers to implement it appears to be tricky.
-
-        So instead for now we do the next best thing and call this
-        class method on all parsers that need to support the repo
-        parameter
-
-        **Arguments**
-
-        - parser (`argparse.ArgParser`)
-        - plugin_config (`dict`)
-        """
-        parser.add_argument(
-            "repository",
-            nargs="?",
-            type=str,
-            help=VersionPluginConfig().repository.help,
-            default=plugin_config.get("repository"),
-        )
 
     @classmethod
     def add_arguments(cls, parser, plugin_config, confu_cli_args):
@@ -176,12 +134,22 @@ class VersionPlugin(ExecutablePlugin):
     def no_auto_dev(self, value):
         self._no_auto_dev = value
 
+    @property
+    def version(self):
+        """ current version as it exists in `version_file` """
+        try:
+            print(("Reading version from", self.version_file))
+            with open(self.version_file) as fh:
+                version = fh.read().strip()
+        except FileNotFoundError:
+            self.log.debug(f"No version file found at {self.version_file}")
+            return "0.0.0"
+        return version
+
     def execute(self, **kwargs):
 
         super().execute(**kwargs)
 
-        branch_dev = self.get_config("branch_dev")
-        branch_release = self.get_config("branch_release")
         self.no_auto_dev = kwargs.get("no_auto_dev", False)
         self.init_version = kwargs.get("init", False)
 
@@ -197,48 +165,6 @@ class VersionPlugin(ExecutablePlugin):
             raise OperationNotExposed(op)
 
         fn(**kwargs)
-
-    def repository(self, target):
-        """
-        Return plugin instance for repository
-
-        **Arguments**
-
-        - target (`str`): name of a configured repository type plugin
-          or filepath to a repository checkout
-
-        **Returns**
-
-        git plugin instance (`GitPlugin`)
-        """
-
-        try:
-            plugin = self.other_plugin(target)
-            if not isinstance(plugin, RepositoryPlugin):
-                raise TypeError(
-                    "The plugin with the name `{}` is not a "
-                    "repository type plugin and cannot be used "
-                    "as a target".format(target)
-                )
-        except KeyError:
-            if target:
-                target = os.path.abspath(target)
-            if not target or not os.path.exists(target):
-                raise OSError(
-                    "Target is neither a configured repository "
-                    "plugin nor a valid file path: "
-                    "{}".format(target)
-                )
-
-            plugin = ctl.plugins.git.temporary_plugin(self.ctl, target, target)
-
-        if not self.init_version and not os.path.exists(plugin.version_file):
-            raise UsageError(
-                "Ctl/VERSION file does not exist. You can set the --init flag to create "
-                "it automatically."
-            )
-
-        return plugin
 
     @expose("ctl.{plugin_name}.merge_release")
     def merge_release(self, repo, **kwargs):
@@ -332,87 +258,3 @@ class VersionPlugin(ExecutablePlugin):
         if not is_dev and not self.no_auto_dev:
             self.log.info("Creating dev tag")
             self.bump(version="dev", repo=repo, **kwargs)
-
-    def update_version_files(self, repo_plugin, version, files):
-
-        """
-        Finds the various files in a repo that will need to
-        have new version values written, such as Ctl/VERSION
-        and pyproject.toml
-        """
-
-        types = ["ctl", "pyproject"]
-
-        for typ in types:
-            fn = getattr(self, f"update_{typ}_version")
-            path = fn(repo_plugin, version)
-            if path:
-                files.append(path)
-
-    def update_ctl_version(self, repo_plugin, version):
-
-        """
-        Writes a new version to the Ctl/VERSION files
-        """
-
-        with open(repo_plugin.version_file, "w") as fh:
-            fh.write(version)
-        return repo_plugin.version_file
-
-    def update_pyproject_version(self, repo_plugin, version):
-
-        """
-        Writes a new version to the pyproject.toml file
-        if it exists
-        """
-
-        pyproject_path = os.path.join(repo_plugin.checkout_path, "pyproject.toml")
-
-        if os.path.exists(pyproject_path):
-            with open(pyproject_path, "r") as fh:
-                pyproject = toml.load(fh)
-
-            pyproject["tool"]["poetry"]["version"] = version
-            with open(pyproject_path, "w") as fh:
-                toml.dump(pyproject, fh)
-            return pyproject_path
-
-    def validate_changelog(self, repo, version, data_file="CHANGELOG.yaml"):
-
-        """
-        Checks for the existance of a changelog data file
-        like CHANGELOG.yaml or CHANGELOG.json and
-        if found will validate that the specified
-        version exists.
-
-        Will raise a KeyError on validation failure
-
-        **Arrguments**
-
-        - version (`str`): tag version (eg. 1.0.0)
-        - repo (`str`): name of existing repository type plugin instance
-        """
-
-        version = version_string(version)
-        repo_plugin = self.repository(repo)
-
-        changelog_path = os.path.join(repo_plugin.checkout_path, data_file)
-
-        if not os.path.exists(changelog_path):
-            return
-
-        changelog_plugin = temporary_changelog_plugin(
-            self.ctl, f"{self.plugin_name}_changelog", data_file=changelog_path
-        )
-
-        self.log.info(f"Found changelog data file at {changelog_path} - validating ...")
-
-        try:
-            changelog_plugin.validate(changelog_path, version)
-        except ChangelogVersionMissing as exc:
-            raise PluginOperationStopped(
-                self,
-                "{}\nYou can set the --no-changelog-validate flag to skip this check".format(
-                    exc
-                ),
-            )
