+++ tests/test_plugin_version.py
@@ -22,15 +22,15 @@ def test_tag(tmpdir, ctlr):
     plugin, dummy_repo = instantiate(tmpdir, ctlr)
     plugin.tag(version="1.0.0", repo="dummy_repo")
     assert os.path.exists(dummy_repo.version_file)
-    assert dummy_repo.version == ("1", "0", "0")
+    assert dummy_repo.version == "1.0.0"
     assert dummy_repo._tag == "1.0.0"
 
     plugin.tag(version="1.0.1", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "0", "1")
+    assert dummy_repo.version == "1.0.1"
     assert dummy_repo._tag == "1.0.1"
 
     plugin.tag(version="1.0.2", repo="dummy_repo", release=True)
-    assert dummy_repo.version == ("1", "0", "2")
+    assert dummy_repo.version == "1.0.2"
     assert dummy_repo._tag == "1.0.2"
     assert dummy_repo._merged == "release"
     assert dummy_repo.branch == "release"
@@ -57,19 +57,19 @@ def test_bump(tmpdir, ctlr):
     plugin.tag(version="1.0.0", repo="dummy_repo")
 
     plugin.bump(version="dev", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "0", "0", "1")
+    assert dummy_repo.version == "1.0.0.1"
     assert dummy_repo._tag == "1.0.0.1"
 
     plugin.bump(version="patch", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "0", "1")
+    assert dummy_repo.version == "1.0.1"
     assert dummy_repo._tag == "1.0.1"
 
     plugin.bump(version="minor", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "1", "0")
+    assert dummy_repo.version == "1.1.0"
     assert dummy_repo._tag == "1.1.0"
 
     plugin.bump(version="major", repo="dummy_repo")
-    assert dummy_repo.version == ("2", "0", "0")
+    assert dummy_repo.version == "2.0.0"
     assert dummy_repo._tag == "2.0.0"
 
     with pytest.raises(ValueError):
@@ -81,22 +81,22 @@ def test_bump_truncated(tmpdir, ctlr):
     plugin.tag(version="1.0", repo="dummy_repo")
 
     plugin.bump(version="minor", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "1", "0")
+    assert dummy_repo.version == "1.1.0"
     assert dummy_repo._tag == "1.1.0"
 
     plugin.tag(version="1.0", repo="dummy_repo")
     plugin.bump(version="patch", repo="dummy_repo")
-    assert dummy_repo.version == ("1", "0", "1")
+    assert dummy_repo.version == "1.0.1"
     assert dummy_repo._tag == "1.0.1"
 
     plugin.tag(version="2", repo="dummy_repo")
     plugin.bump(version="patch", repo="dummy_repo")
-    assert dummy_repo.version == ("2", "0", "1")
+    assert dummy_repo.version == "2.0.1"
     assert dummy_repo._tag == "2.0.1"
 
     plugin.tag(version="3", repo="dummy_repo")
     plugin.bump(version="major", repo="dummy_repo")
-    assert dummy_repo.version == ("4", "0", "0")
+    assert dummy_repo.version == "4.0.0"
     assert dummy_repo._tag == "4.0.0"
